aodh-2018-backend/requirements.txt
boto==2.48.0
boto3==1.7.45
botocore==1.10.45
bz2file==0.98
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dill==0.2.8.2
docutils==0.14
ekphrasis==0.4.10
flasgger==0.9.0
Flask==1.0.2
Flask-Cors==3.0.6
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gensim==3.4.0
google-api-core==0.1.4
google-auth==1.5.0
google-cloud==0.33.1
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.2
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
googlemaps==3.0.1
grpc-google-iam-v1==0.11.4
grpcio==1.12.1
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
jaconv==0.2.3
JapaneseTokenizer==1.3.7
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
many-stop-words==0.2.2
MarkupSafe==1.0
mecab-python3==0.7
mistune==0.8.3
nltk==3.3
numpy==1.14.5
oauth2client==3.0.0
oauthlib==2.1.0
pexpect==4.6.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.6.0
psutil==5.4.6
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pypandoc==1.4
python-dateutil==2.7.3
python-twitter==3.4.2
pytz==2018.4
PyYAML==3.12
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.5.7
typing==3.6.4
ujson==1.35
urllib3
Werkzeug==0.14.1
yweather==0.1.1
